put administration view functions in admin_views (own package), code view for teachers to manage their courses (def teacher in views)
